pyproject.toml
[project]
name = "flaskproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-jwt-extended>=4.7.1",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-security>=5.6.2",
    "flask-security-too>=5.6.2",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "werkzeug>=3.1.3",
]
